@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = "1.2.71"
+    ext.kotlin_version = "1.3.21"
     repositories {
         google()
         jcenter()
@@ -16,10 +16,10 @@ buildscript {
         //-- Fix end
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0'
+        classpath "com.android.tools.build:gradle:3.3.2"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // For Firebase
-        classpath 'com.google.gms:google-services:4.1.0'
+        classpath "com.google.gms:google-services:4.2.0"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
