@@ -2,7 +2,7 @@ package com.flyingkite.mytoswiki.dialog;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.LinearLayout;
 
