@@ -6,7 +6,7 @@ import java.util.Map;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.LinearLayout;
 import com.flyingkite.fabric.FabricAnswers;
