@@ -2,7 +2,7 @@ package com.flyingkite.mytoswiki.tos;
 
 import android.content.Context;
 import android.content.res.AssetManager;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import com.flyingkite.crashlytics.CrashReport;
