@@ -1,6 +1,6 @@
 package com.flyingkite.util;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.flyingkite.library.log.Loggable;
 
